+++ app/Http/Controllers/UsersController.php
@@ -18,6 +18,10 @@ class UsersController extends Controller {
 	 */
 	public function index()
 	{
+		//TODO fix (Authentication doesn't work with view_users?)
+		//if ( !Auth::user()->can('view_users') )
+		//	return response('Unauthorised', 403);
+
 		$users = User::all();
 
 		return view('users.index')->with('users', $users);
@@ -30,8 +34,8 @@ class UsersController extends Controller {
 	 */
 	public function create()
 	{
-		//if ( !Auth::user()->can('add_users') )
-		//	return response('Unauthorised', 401); //TODO Move to routes?
+		if ( !Auth::user()->can('add_users') )
+			return response('Unauthorised', 403);
 
 		$roles = Role::lists('display_name', 'name');
 
@@ -45,6 +49,9 @@ class UsersController extends Controller {
 	 */
 	public function store(Request $request)
 	{
+		if ( !Auth::user()->can('add_users') || !Auth::user()->can('assign_roles') )
+			return response('Unauthorised', 403);
+
 		//Make sure the password match
 		if( $request->input('password') !== $request->input('password2') )
 			return redirect()->back(); //TODO add message
@@ -54,6 +61,10 @@ class UsersController extends Controller {
 			'email' => $request->input('email'),
 			'password' => bcrypt($request->input('password'))
 		));
+
+		//Assign user role
+		$role = Role::where('name', '=', $request->input('role'))->first();
+		$user->attachRole($role->id);
 
 		//Run the index method to display all the users
 		return redirect()->route('users.index');
@@ -78,6 +89,9 @@ class UsersController extends Controller {
 	 */
 	public function edit($id)
 	{
+		if ( Auth::user()->id != $id && !Auth::user()->can('edit_users') )
+			return response('Unauthorised', 403);
+
 		$user  = User::FindorFail($id);
 		$roles = Role::lists('display_name', 'name');
 
@@ -92,6 +106,9 @@ class UsersController extends Controller {
 	 */
 	public function update(Request $request, $id)
 	{
+		if ( Auth::user()->id != $id && !Auth::user()->can('edit_users') )
+			return response('Unauthorised', 403);
+
 		//Setup passwords as variables to check if empty
 		$pass 	= $request->input('password');
 		$pass2 	= $request->input('password2');
@@ -108,8 +125,8 @@ class UsersController extends Controller {
 		if( ( !empty( $pass ) && !empty($pass2) ) && $request->input('password') == $request->input('password2') )
 			$user->password = bcrypt($request->input('password'));
 
-		//TODO check permissions
-		if( true ){
+		//If the user cna assign roles w cn update the role with the post data
+		if( Auth::user()->can('assign_roles') ){
 			//Remove all existing roles
 			$user->roles()->detach();
 
@@ -132,6 +149,9 @@ class UsersController extends Controller {
 	 */
 	public function destroy($id)
 	{
+		if ( !Auth::user()->can('delete_users') )
+			return response('Unauthorised', 403);
+
 		$user = User::FindOrFail($id)->delete();
 
 		//Run the index
